feat: Add BitCoinExchange program
This commit introduces a new program called BitCoinExchange. The program
reads data from a CSV file and stores it in a map. The keys of the map are
dates in the format "YYYY-MM-DD", and the values are floats. The program
also handles file opening errors and prints an error message if it fails
to open the file.

The Makefile has been updated to change the name of the executable from
'exec' to 'btc'. The 'run' rule has been removed from the Makefile.

A test CSV file has been added for testing purposes.